+++ CHANGELOG.md
@@ -8,6 +8,7 @@ We follow [Semantic Versions](https://semver.org/).
 - Add `NestedFieldsMixin` for serializers
 - Add ability to add extra kwargs for ordering fields in `OrderingFilterBackend`.
 Also add new settings `SARITASA_DRF_ORDERING_IS_NULL_FIRST` and `SARITASA_DRF_ORDERING_IS_NULL_LAST`.
+- Add ability serialize multiple objects in `ApiActionTester.serialize_data` method via `many=True`.
 
 ## 0.1.0